backend/requirements.txt
fastapi==0.115.13
pydantic==2.11.5
feedparser==6.0.11
PyPDF2==3.0.1
python-dotenv==1.1.0
openai==1.86.0
uvicorn==0.34.3
requests==2.32.4